requirements.txt
flask==2.1.0
flask-cors==3.0.10
pymongo==4.6.1
scikit-learn==1.2.2
pandas==1.5.3
python-dotenv==1.0.0
gunicorn==20.1.0
nltk==3.7
dnspython==2.4.2
spacy==3.5.2
matplotlib==3.7.1
scipy==1.10.1